Move UI modules into library
