+++ lib/ui/gauge.cpp
