+++ lib/ui/gauge.h
