+++ lib/ui/gauge_anim.cpp
