+++ lib/ui/gauge_anim.h
