+++ lib/ui/gauge_tile.cpp
