+++ lib/ui/gauge_tile.h
