+++ lib/ui/voice_synth.cpp
