+++ lib/ui/voice_synth.h
